+++ packages/engine/Source/Scene/GlobeSurfaceTileProvider.js
@@ -2355,7 +2355,6 @@ const surfaceShaderSetOptionsScratch = {
   hasGeodeticSurfaceNormals: undefined,
   hasExaggeration: undefined,
   vectorAntialias: undefined,
-  vectorWidthInMeters: undefined,
 };
 
 const defaultUndergroundColor = Color.TRANSPARENT;
@@ -2611,8 +2610,6 @@ function addDrawCommandsForTile(tileProvider, tile, frameState) {
   surfaceShaderSetOptions.hasExaggeration = hasExaggeration;
   surfaceShaderSetOptions.vectorAntialias =
     tileProvider.vectorProvider.antialias;
-  surfaceShaderSetOptions.vectorWidthInMeters =
-    tileProvider.vectorProvider.widthInMeters;
 
   const tileImageryCollection = surfaceTile.imagery;
   let imageryIndex = 0;